Make error handling more concise with static str
I think this is a Bad error type, because it can't be easily used
programmatically, but that's fine since this is intended for a CLI (at least
for now).

Also it seems a bit insane to require a static str but actually I think that's
a pretty reasonable restriction for these error messages.

I dunno maybe I've written too much Go and I'm poisoned by pragmatism.