Clean up type for revspecs a bit
This is still a bit messy but at least we have a kinda commentary to declare
that this OsString is supposed to have the given meaning.

This also requires cleaning up some mess where I used String instead of
OsString for these values (similar mess remains for the args and program and
stuff).