New attempt at higher-level command-running API
This feels a little more granular and works on both async and sync command
APIs. It's perhaps still a bit awkward that calling ok() breaks your method
chain because it returns () - I'm not sure how best to work around that given
that you can't just return Self from a trait method. But anyway, if you have a
nontrivial method chain (as is actually the case in Repo::rev_list), this isn't
the method for you anyway...?