Noodling around with worktree API
The API here for temporary worktrees was garbage temp worktrees had a special
public API instead of just being a very similar object that happened to have a
side effect on drop.

This is an attempt to unify them but I am not really sure this is The Way, it
feels a bit like I'm  basically writing C++ in Rust. Maybe "a similar object
that happens to have a side effect on drop" is the key here: perhaps the right
design is to have two separate types, and have one Borrow as the other or
something, and then just write code that is generic across the two types? What
we have here looks wacky and inheritancey.

While noodling with thisI also found I had to rework the argument to
Manager::new, this was bad before so now I've just switched it to something
else bad. I think I need to read
[this](https://github.com/pretzelhammer/rust-blog/blob/master/posts/common-rust-lifetime-misconceptions.md#2-if-t-static-then-t-must-be-valid-for-the-entire-program)
and maybe I'll start to understand the 'static thing a bit better? But right
now I'm tired again (why am I always so tired?).

Anyway, the plan is to totally rework the Manager API anyway (this is a
learning project, flagrant refactoring is part of the point), with async code
we do not actually need this worker+queue architecture, we should instead just
have a pool of workqueues and spin up tasks immediately that just block on this
pool. I think this refactoring will be instructive too.

After that, time to check of some more TODOs then we can start looking at
implementing an actual API to grab results back from the Manager.