pool: Implement Deref instead of AsRef
I was a bit thrown off because I thought that Deref needed to somehow access
the underlying object in an owned way?? But it doesn't, it just kinda converts
to a reference.

It feels like the main difference between AsRef and Deref is that Deref has an
associated type, meaning you can only implement it for one type.

So AsRef is like "I can be converted to some other types" whereas Deref is like
"I AM a reference to this other type".