Add a settle method to the test::Manager
Idea here is to be able to detect when we have received all the events we're
gonna receive, without having to wait an arbitrary amount of time in tests.

This is hacky as fuck! I dunno if I'm being stupid here. Maybe instead I should
have found a way to join the tasks directly instead of this nonsense with a
counter.