Fix checking for double-allocated worktrees
Idea here is to create a "lockfile" to mark the worktree as "used", then signal
via creation of a secondary file if the presence of this file indicates two
jobs used the same worktree at once.

But, I put the lockfile in the shared temp directory instead of a per-worktree
location.