Simplify JobCounter using tokio's watch
As Konstantin pointed out and demonstrated in
https://github.com/Mandragorian/local-ci/commit/d8bf91a74f89c50730c63d6aca41c3512e5bc05f,
watch is more powerful than I thought, and in this case we can basically just
use it as a condition variable.

I wrote this out from scratch to see if it would look the same as Konstantin's
(I had read it but didn't pay any attention to the details) and what I ended up
with is exactly the same except that I used send_if_modified instead of
send_modify.