RESOURCE THROTTLING MEGA COMMIT
This replaces the sensible and pleasing Pool<T> type with a Pools<T> type,
which supports a single pool of type T, plus an arbitrary number of "token
pools", which don't directly track resources. This is really ugly but it
satisfies our specific needs right here.

These token pools are then used in turn to let Tests specify which and how many
of these tokens they need.