Fix should_not_settle flakes (hack)
For some reason I sometimes see the `read` in this test script return (I can
verify this by looking in the logs in the output database). There is one good
reason why this should happen which is described in the big comment I added.
I'm not sure if that's really why.

Anyway, if we do add it, and replace the read with a sleep infinity (necessary
for the reason also described in the comment), the issue goes away.