Port integration_test to async
Since I wanna use the Git library. But also I think just life will be less
confusing if things are async throughout.

Note the tokio process API has a builtin kill-on-drop functionality, presumably
this is exactly because it's annoying to do that work, which is async, from a
drop function which is sync.