Store reference to whole Test from TestResult
Not sure if this is a  mistake...? It's a little tricky because until now the
TestCase was a kinda "plain only data" thing that made obvious sense to use as
a map key or something.

We only actually do that in the tests though. Maybe for the moment it's fine.
If we need to do it IRL we should probably define some custom hashing method or
something? I dunno.