Break up test::Manager::set_revisions
This requires rejigging the relationships between data structures a bit. This
leads to a lot of code touching a lot of structures' fields, I dunno if this is
a good idea.

Overall these data structures feel like a bit of a mess, I've never really sat
down and tried to think them through very well.