Add test for script that fails when cancelled
Er this test code is pretty messy, but I can't be bothered to rework it.

I expected this test to fail because I believe I'm seeing failure test results
get cached when I shutdown local-ci. But the test passes, and now I realise
that the task spawned in spawn_job does indeed skip storing the result in the
database if it observes TestStatus::Canceled.

Presumably then, the bug is actually specific to the case that the jobs inherit
the SIGINT, and this means they shut down themselves and we get a
TestStatus::Completed(errorcode) before we cancel the job ourself.