Always construct test::Test from test::Config
In a coming commit I'll make test::Test a hashable type, but because I think
I'm a very clever boy with clever ideas who loves to make things needlessly
complicated, I'm gonna use the config hash as the hash of the Test value
itself. So it will become invalid to construct a Test directly because its
config hash will be wrong (that is already the case, but currently it's
harmless because the config hash is only used to index the result database).