Revert "Switch test::Manager::job_cts to be indexed by TestCase"
This reverts commit 6b503bc3dc537550d086d60b39475948ad16656a.

This implementation was dumb. Making the TestCase Hash was stupid
because it requires the Test to be Hash and Eq. I implemented Eq by just
comparing hashes but that is not how hashing works, it can have
collisions. We need to actually compare all the fields, this feels
inefficient (I don't care in reality but I'm supposed to be learning
Rust here, not learning how to write Python in a more complex syntax).