Treat worktrees less specially as a resource
Hmm, not super happy with how this turned out, but it does feel like a small step in the right direction.

The ultimate goal here is:

 - It should not be required for tests to request a worktree
 - Tests should be able to request resources that have a "value" which will be fed to them as an env var.

So we try to get rid of the "obj" thing in the resource pools API a little bit
and just have the items returned by the pool be an enum variant, which the user
is supposed to trust will match the variant of the key they provided.