Clean up non-atomic file-op flakes
I think I asked ChatGPT or trusted some Stack Overflow answer that says when
you write a file in bash it appears atomically with content in place. I don't
think that is actually the case, I'm seeing failures where the PID file is
empty.

Clean up those failures, and fix the bug by creating the file via a mv
operation.

This means that we lose the ability to count runs which we were hackily getting
from the PID file. So create a new file for just that purpose.