Make ID an associated type of GraphNode
This is a bit more concise, and also should let me try to implement this logic
using an iterator, to unify hte code  a bit.

Potentially this is dumb because it means any given type can only be a node in
a single type of graph? I dunno, don't think I care.