Fix SIGINT handling in test script (COMMIT MESSAGE WIP)
You cannot have multiple traps for the same signal. This code mostly gets away
with it, because we rarely rely on SIGINT working in the same cases that we
also expect the exit code commit-message-tag logic to work.

This is not the bug that I'm trying to debug, but it does seem like it is _a_
bug. So, try to clean that up a bit.