Mega commit: Empower rev_parse to make TestCase construction cleaner
At the moment we construct TestCases by passing in a repo, and the constructor
uses that to look up a tree hash where that's the appropriate key for storing
the result in the database.

This is dumb because a) why the hell does a TestCase (supposed to basically
just be a tuple) have to care about doing Git operations b) it puts async logic
into annoying places to have async logic and c) it prevents you from
effectively re-using tree lookups.

So, this creates a slightly higher-level type to represent a Commit, where we
can also store the tree, since this is basically free to lookup at the time
when we call rev_parse.

Now we don't wanna _entirely_ rewrite everything to use this new type, but life
is easier if we do have the commit method return it, so this commit does get
pretty big in order to deal with that.

As a workaround for not wanting everything to have to use the new type,
set_revisions becomes generic over both CommitHash and Commit. This makes
calling it with empty collections kinda annoying because the compiler wants to
know what type that empty collection would contain, so we create a new helper
called cancel_running for that usecase.