Don't leak worktrees if the TempWorktree constructor is cancelled
Have I tested this? No, fuck off. I plan for cancelling worktree creation to
become normal when I add the test command, since we wanna kick off worktree
setup immediately and in parallel, then kick off jobs and have them pick from
worktrees as they become available. This means that if jobs run much quicker
than worktree creation we will get a faster turnaround.

Actually... do I need any of that shit? Probably not. But anyway, it's nice not
to leak worktrees.

But, this work in drop is shit. Would be nice to get rid of that I guess.