WORKTREE CREATION REFACTOR MEGA COMMIT
In order to implement the test command, we need to reduce the godscope of the
test manager a bit. Here we're moving out the creation of worktrees.

In order to be b l a z i n g  f a s t we do this in an overoptimised way where
we set them up in parallel with the actual startup of the manager & jobs, this
means the first job can start as soon as a worktree is ready; and if jobs are
much faster than worktree setup (which is plausible for bigg reppos) we can
even run jobs before the second worktree is ready.

The hellish nightmare part of this is cleanup: we wanna do the worktree
teardown in an async context so we need to call cleanup explicitly. Rust...
doesn't have a solution for this. We just have to fucking not return from the
function until we've done that.