Better API for the text rendering lib
fmt::Write doesn't allow propagating the error from the underlying output. (It
propagates that an error occurred, but not what it was). This gave me the
feeling I was abusing it here. I did a bit of googling and found this:

https://users.rust-lang.org/t/std-write-and-std-write-interopability/110614/2

Which made me realise that we can achieve the same kind of "minimal copying"
properties of the `&mut fmt::Write` design by just returning a struct that
implements Display. Now the core stdlib APIs provide this shim from io::Write
to fmt::Write that I was looking for.

Noice.