Switch up text styling
I wanna have some more flexibility here for the UI, I think this will be better
if I use a CSS stylesheet after all.

So now instead of direct control of the classes the user of the API just sets a
class. This naturally led to some rejigging of the text styling API. That means
it no longer looks like the ratatui API. I guess that's fine.

This also actually changes the UI styling a bit, the colours were kinda wack.