refactor: Fiddle around with Job API
This API is pretty confusing, and it seems to have a bug: it doesn't actually
take care of storing stuff in the database or sending completion notifications.

In order to take care of that, it will need to construct the TestStatus, so now
we'll have to haveit return that instead of an anyhow::Result. In order to make
that work better we provide a From to convert to a Result if that's what you
wanted.