fix: Move test result reporting into Job
At present there is some logic in Manager::spawn_job that other code expects
the Job to take care of. So, make it take care of it.

IIUC this is not the only bug here - we need basically all the logic to be
moved. But just going one step at a time.