refactor: Move output creation into TestJob::run
This whole part of the API is a total mess but I can't quite figure it out and
I want to get locking set up.

Here's an attempt to at least fix the issue that we construct `JobOutput`s
before checking if there is an existing entry in the database. The weird Job
API is not fixed.