TIRED AGAIN - slightly better attempt at locking (NOTES IN MSG)
Brain stopped working now, but I think what we want to do now is kinda
explained in the comment. We need to open the JSON file (creating it if it
doesn't exist), then lock it exclusively. Once we have the lock we should be
able to figure out if someone already ran it. If they did we would downgrade
this to a non-exclusive lock to prevent it from being deleted while we're using
it (but this is step 2/3 of the problem). If they didn't then we wanna hold the
lock until we're done producing output.