refactor: Split up TestStatus and TestOutcome
It's confusing that there is one type that mashes together final statuses and
ongoing statuses.

I seem to remember flailing around a bit with this but not really feeling
bothered to try and design it properly. Well, now I'm trying to clean up the
TestJob interface a bit I think it's time to clean it up slightly.