refactor: Clean up TestJob notifying etc a bit
Now that the TestOutcome is a Result and there's less stupidity around error
handling, the responsibilities of the different methods can be clarified a
little bit.

Now there's still this wack run_with thing and the awkward fact that run
returns the outcome even though it already did the notifying for you.

But at least now the notify_completion calls are in a sane place, i.e. exactly
one per "run" entry point.