prep: Create DatabaseOutput::ephemeral
This is more preparation for allowing set_result to return the database entry
it created. At present this is painful because of the polymorphism at play.
Simplify things by removing polymorphism and instead just having a variant of
the database entry that isn't actually in the database (but is otherwise just
the same thing).