prep: Make TestOutcome contain a DB entry
The fact that this created such an enormous pain for the tests might be a sign
it's just a bad design. Until now the TestOutcome has been inert data. I want
to pass the locked database entry through to the dependencies though.

I think the right thing to do might have been to have a separate mechanism to
do that, or do it via an outer wrapper for the existing inert data :/

Anyway, I don't wanna faff around refactoring this forever so here's something
that works.