cleanup: Avoid N git commands for building UI buffer
Per
https://github.com/bjackman/limmat/issues/17#issuecomment-2561479484
this seems to be one reason why Limmat can be unresponsive when
running on big repos, especially with crippled Git.

I don't think this will make the issue go away but it does seem worth
avoiding zillions of git commands anyway, this probably has some
tangible benefit even if the system is architected to isolate
the performance impact of those commands.