test: Shitty bullshit hacks for printf debugging in shitty tests
fukken shitty fuekken grumble fuckin shit rust is shit

The test-logger create is crap, it doesn't seem to have a reasonable way to add
timestamps to the logs.

Rust's test harness is also crap, there's no way to configure tests to do
something special before they run. Look at this shit:

https://stackoverflow.com/questions/58006033/how-to-run-setup-code-before-any-tests-run-in-rust

https://stackoverflow.com/questions/67087597/is-it-possible-to-use-rusts-log-info-for-tests

My plan for a moment was to swtich to ctor and use it to just manually
ininitalize the flexi_logger crate but that is also buggy, Rust says the stdlib
is not guaranteed to work before main so that's also just total shite.

If I'm gonna be sitting in a room that stinks of shit, at least let it be my
_own_ shit that stinks.