fix: Don't hold database locks from UI code
Fixes https://github.com/bjackman/limmat/issues/24.

The proper way to do this would be to have a proper way to remember the
abstracted result of a test case without having a reference to the database
entry. But, I really don't feel like it.

A super shitty lazy solution is just to render the text immediately and store
that. And this is just done with clone-crazy laziness.

(cherry picked from commit 91e210f6f84fb64e35f353991801d9e09d4caec6)